extract aggregator
